Move SegmentedControl to from-hdf-to-asff-mapper, move all reverse-asff mapper files into one folder, clean up evaluate()
