+++ libs/hdf-converters/test/hdf_checker.spec.ts
@@ -3,8 +3,8 @@ import {ExecJSON} from 'inspecjs';
 import _ from 'lodash';
 import {ASFFMapper} from '../src/asff-mapper';
 import {BurpSuiteMapper} from '../src/burpsuite-mapper';
-import {IFindingASFF} from '../src/converters-from-hdf/asff-types';
-import {FromHdfToAsffMapper} from '../src/converters-from-hdf/from-hdf-to-asff-mapper';
+import {IFindingASFF} from '../src/converters-from-hdf/asff/asff-types';
+import {FromHdfToAsffMapper} from '../src/converters-from-hdf/asff/reverse-asff-mapper';
 import {JfrogXrayMapper} from '../src/jfrog-xray-mapper';
 import {NiktoMapper} from '../src/nikto-mapper';
 import {SarifMapper} from '../src/sarif-mapper';
